Adds an autotune component to base character
This fixes the issue where only squad net was set
but not platoon net. When a player spawns or
picks up a radio we check their faction and group
information to get the radio freqs for squad and
platoon. The first channel is always set to squad
and the second channel is always platoon.